step 09: separate classes
