+++ firebase-get-to-know-flutter/step_09/lib/guest_book_message.dart
@@ -1,0 +1,6 @@
+class GuestBookMessage {
+  GuestBookMessage({required this.name, required this.message});
+
+  final String name;
+  final String message;
+}
